CMSPixel: change behavior when reading multiple I2C devices
if no "i2c" parameter is given, the address defaults to "0" and
the config files are read as written in the config file. If "i2c"
is given, the file path is appended with the "_Cx" pattern, where
x is the i2c address. This works now for both single chips and
multiple ROCs.

+++ producers/cmspixel/include/CMSPixelProducer.hh
@@ -34,8 +34,8 @@ private:
 
   // Helper function to read DACs from file which is provided via eudaq config:
   std::vector<std::pair<std::string,uint8_t> > GetConfDACs(int16_t i2c = -1);
-  std::vector<uint8_t> &split(const std::string &s, char delim, std::vector<uint8_t> &elems);
-  std::vector<uint8_t> split(const std::string &s, char delim);
+  std::vector<int32_t> &split(const std::string &s, char delim, std::vector<int32_t> &elems);
+  std::vector<int32_t> split(const std::string &s, char delim);
 
   std::vector<pxar::pixelConfig> GetConfTrimming(int16_t i2c = -1);
